apps/api/pyproject.toml
[project]
name = "human-ai-interaction-backend"
version = "0.1.0"
description = "This project contains the backend for the application we develop in the 'Human-AI Interaction' course."
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi>=0.121.2",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.35",
    "aiosqlite>=0.20.0",
    "pydantic-settings>=2.5.2",
    "bcrypt>=4.2.0",
    "pyjwt>=2.9.0",
    "greenlet>=3.2.4",
]

[dependency-groups]
test = [
    "pytest",
    "pytest-asyncio>=0.23.7",
    "httpx>=0.27.2",
]
lint = [
    "ruff",
]
typing = [
    "mypy"
]
dev = [
    { include-group = "test" },
    { include-group = "lint" },
    { include-group = "typing" },
]
